# docs/swagger.yaml
basePath: /api-market-testnet
definitions:
...
  request.BroadcastTxReq:
    properties:
      txHex:
        type: string
    type: object
  request.BuildIdCoinsCommitRequest:
    properties:
      commitTxOutIndex:
        type: integer
      commitTxRaw:
        type: string
      orderId:
        type: string
    type: object
...
  request.ColdDownDummyUtxoRequest:
    properties:
      address:
        type: string
      amount:
        type: integer
      changeAddress:
        type: string
      count:
        type: integer
      feeRate:
        type: integer
      index:
        type: integer
      perAmount:
        type: integer
      pkScript:
        type: string
      priKeyHex:
        type: string
      txId:
        type: string
      utxoType:
        $ref: '#/definitions/models.UtxoType'
    type: object
...
  request.TestAuthReq:
    properties:
      address:
        type: string
    type: object
...
  respond.BroadcastTxResp:
    properties:
      txId:
        type: string
    type: object
...
  respond.FetchIdCoinsDeployCheckResp:
    properties:
      canDeploy:
        type: boolean
      msg:
        type: string
    type: object
  respond.FetchIdCoinsListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/respond.IdCoinsInfoResp'
        type: array
      total:
        type: integer
    type: object
...
  respond.Message:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      processingTime:
        type: integer
    type: object
...
  respond.Mrc20BalanceInfoResp:
    properties:
      list:
        items:
          $ref: '#/definitions/respond.Mrc20BalanceInfo'
        type: array
      total:
        type: integer
    type: object
...
  respond.Mrc20ShovelResp:
    properties:
      list:
        items:
          $ref: '#/definitions/respond.ShovelInfo'
        type: array
      total:
        type: integer
    type: object
...
  respond.Mrc20TickListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/respond.Mrc20TickInfo'
        type: array
      total:
        type: integer
    type: object
  respond.Mrc20TickMarketPriceResp:
    properties:
      decimals:
        type: integer
      floorPrice:
        type: string
      floorPriceUsd:
        type: string
      lastPrice:
        type: string
      lastPriceUsd:
        type: string
      marketCap:
        type: string
      marketCapUsd:
        type: string
      price:
        type: string
      priceUsd:
        type: string
      supply:
        type: string
      tick:
        type: string
      tickId:
        type: string
      tokenName:
        type: string
      totalVolume:
        type: integer
    type: object
...
  respond.Mrc20UtxoResp:
    properties:
      list:
        items:
          $ref: '#/definitions/respond.Mrc20Utxo'
        type: array
      total:
        type: integer
    type: object
...
  respond.TakeOrderResp:
    properties:
      assetId:
        type: string
      assetType:
        $ref: '#/definitions/models.AssetType'
      orderId:
        type: string
      orderState:
        $ref: '#/definitions/models.OrderState'
      txId:
        type: string
    type: object
...
info:
  contact:
    name: API Support
  description: MetaID-Market API Service
  title: MetaID-Market API Service
  version: "1.0"
paths:
  /api/v1/admin/utxo/colddown:
    post:
      description: Colddown Market utxo
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.ColdDownDummyUtxoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Message'
      summary: Colddown Market utxo
      tags:
      - Market-Admin
  /api/v1/auth/test:
    post:
      description: Take Market order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.TestAuthReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.TakeOrderResp'
      summary: Take Market order
      tags:
      - Market-Order
  /api/v1/common/mrc20/address/balance-list:
    get:
      description: Fetch address mrc20 balances
      parameters:
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: cursor
        in: query
        name: cursor
        type: integer
      - description: size：max=50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20BalanceInfoResp'
      summary: Fetch address mrc20 balances
      tags:
      - Common
  /api/v1/common/mrc20/address/shovel:
    get:
      description: Fetch address mrc20 shovels
      parameters:
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: tickId
        in: query
        name: tickId
        required: true
        type: string
      - description: cursor
        in: query
        name: cursor
        type: integer
      - description: size：max=50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20ShovelResp'
      summary: Fetch address mrc20 shovels
      tags:
      - Common
  /api/v1/common/mrc20/address/utxo:
    get:
      description: Fetch address mrc20 utxo
      parameters:
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: tickId
        in: query
        name: tickId
        required: true
        type: string
      - description: cursor
        in: query
        name: cursor
        type: integer
      - description: size：max=50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20UtxoResp'
      summary: Fetch address mrc20 utxo
      tags:
      - Common
  /api/v1/common/mrc20/id-coins/address/utxo:
    get:
      description: Fetch address mrc20 id-coins utxo
      parameters:
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: tickId
        in: query
        name: tickId
        type: string
      - description: cursor
        in: query
        name: cursor
        type: integer
      - description: size：max=50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20UtxoResp'
      summary: Fetch address mrc20 id-coins utxo
      tags:
      - Common
  /api/v1/common/mrc20/market/price/info:
    get:
      description: Fetch mrc20 market price info
      parameters:
      - description: tickId
        in: query
        name: tickId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20TickMarketPriceResp'
      summary: Fetch mrc20 market price info
      tags:
      - Common
  /api/v1/common/mrc20/tick/info:
    get:
      description: Fetch mrc20 tick info
      parameters:
      - description: tickId
        in: query
        name: tickId
        type: string
      - description: tick
        in: query
        name: tick
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20TickInfo'
      summary: Fetch mrc20 tick info
      tags:
      - Common
  /api/v1/common/mrc20/tick/info-list:
    get:
      description: Fetch mrc20 tick info list
      parameters:
      - description: pinnumber/totalminted/holders/txcount/change24H/lastPrice/marketCap/totalSupply
        in: query
        name: orderBy
        type: string
      - description: searchTick
        in: query
        name: searchTick
        type: string
      - description: -1/1, if orderBy is 'change24H/lastPrice/marketCap', it works,
          default '-1'
        in: query
        name: sortType
        type: integer
      - description: true/false/null, default null
        in: query
        name: completed
        type: boolean
      - description: cursor
        in: query
        name: cursor
        type: integer
      - description: size：max=50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20TickListResp'
      summary: Fetch mrc20 tick info list
      tags:
      - Common
  /api/v1/common/tx/broadcast:
    post:
      description: Broadcast Tx
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.BroadcastTxReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.BroadcastTxResp'
      summary: Broadcast Tx
      tags:
      - Common
  /api/v1/id-coins/address/mint/order:
    get:
      description: Fetch address id-coins mint orders
      parameters:
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: tickId
        in: query
        name: tickId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.FetchOneIdCoinsMintOrderResp'
      summary: Fetch address id-coins mint orders
      tags:
      - IdCoins
  /api/v1/id-coins/coins-info:
    get:
      description: Fetch id-coins info
      parameters:
      - description: tickId
        in: query
        name: tickId
        required: true
        type: string
      - description: tick
        in: query
        name: tick
        type: string
      - description: issuerAddress
        in: query
        name: issuerAddress
        type: string
      - description: address
        in: query
        name: address
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.IdCoinsInfoResp'
      summary: Fetch id-coins info
      tags:
      - IdCoins
  /api/v1/id-coins/coins-list:
    get:
      description: Fetch id-coins info list
      parameters:
      - description: address
        in: query
        name: address
        type: string
      - description: searchTick
        in: query
        name: searchTick
        type: string
      - description: 'supply/pool/timestamp, default:timestamp  '
        in: query
        name: orderBy
        type: string
      - description: default:-1
        in: query
        name: sortType
        type: integer
      - description: followerAddress
        in: query
        name: followerAddress
        type: string
      - description: default:0
        in: query
        name: cursor
        type: integer
      - description: default:10, max:50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.FetchIdCoinsListResp'
      summary: Fetch id-coins info list
      tags:
      - IdCoins
  /api/v1/id-coins/deploy/check/info:
    get:
      description: Fetch id-coins info
      parameters:
      - description: address
        in: query
        name: address
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.FetchIdCoinsDeployCheckResp'
      summary: Check id-coins deploy check info
      tags:
      - IdCoins
  /api/v1/id-coins/deploy/commit:
    post:
      description: Deploy id coins commit
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.BuildIdCoinsCommitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.BuildIdCoinsCommitResp'
      summary: Deploy id coins commit
      tags:
      - IdCoins
  /api/v1/id-coins/deploy/pre:
    post:
      description: Deploy id coins pre
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.BuildIdCoinsPreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.BuildIdCoinsPreResp'
      summary: Deploy id coins pre
      tags:
      - IdCoins
  /api/v1/id-coins/inscribe/orders:
    get:
      description: Fetch inscribe id-coins orders
      parameters:
      - description: default:deploy, mint
        in: query
        name: opOrderType
        required: true
        type: string
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: tickId
        in: query
        name: tickId
        type: string
      - description: default:0
        in: query
        name: cursor
        type: integer
      - description: default:10, max:50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.FetchIdCoinsOpOrdersResp'
      summary: Fetch inscribe id-coins orders
      tags:
      - IdCoins
  /api/v1/id-coins/mint/commit:
    post:
      description: Mint id coins commit
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.IdCoinsMintCommitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.IdCoinsMintCommitResp'
      summary: Mint id coins commit
      tags:
      - IdCoins
  /api/v1/id-coins/mint/pre:
    post:
      description: Mint id coins pre
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.IdCoinsMintPreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.IdCoinsMintPreResp'
      summary: Mint id coins pre
      tags:
      - IdCoins
  /api/v1/inscribe/mrc20/deploy/commit:
    post:
      description: Inscribe Mrc20 deploy commit
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.Mrc20DeployCommitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20DeployCommitResp'
      summary: Inscribe Mrc20 deploy commit
      tags:
      - Inscribe-Mrc20
  /api/v1/inscribe/mrc20/deploy/pre:
    post:
      description: Inscribe Mrc20 deploy pre
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.Mrc20DeployPreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20DeployPreResp'
      summary: Inscribe Mrc20 deploy pre
      tags:
      - Inscribe-Mrc20
  /api/v1/inscribe/mrc20/mint/commit:
    post:
      description: Inscribe Mrc20 Mint Commit
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.Mrc20MintCommitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20MintCommitResp'
      summary: Inscribe Mrc20 Mint Commit
      tags:
      - Inscribe-Mrc20
  /api/v1/inscribe/mrc20/mint/pre:
    post:
      description: Inscribe Mrc20 Mint Pre
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.Mrc20MintPreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20MintPreResp'
      summary: Inscribe Mrc20 Mint Pre
      tags:
      - Inscribe-Mrc20
  /api/v1/inscribe/mrc20/orders:
    get:
      description: Fetch inscribe mrc20 orders
      parameters:
      - description: default:deploy, mint, transfer
        in: query
        name: opOrderType
        required: true
        type: string
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: tickId
        in: query
        name: tickId
        type: string
      - description: default:0
        in: query
        name: cursor
        type: integer
      - description: default:10, max:50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.FetchMrc20OpOrdersResp'
      summary: Fetch inscribe mrc20 orders
      tags:
      - Inscribe-Mrc20
  /api/v1/inscribe/mrc20/transfer/commit:
    post:
      description: Inscribe Mrc20 Transfer Commit
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.Mrc20TransferCommitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20TransferCommitResp'
      summary: Inscribe Mrc20 Transfer Commit
      tags:
      - Inscribe-Mrc20
  /api/v1/inscribe/mrc20/transfer/pre:
    post:
      description: Inscribe Mrc20 Transfer Pre
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.Mrc20TransferPreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20TransferPreResp'
      summary: Inscribe Mrc20 Transfer Pre
      tags:
      - Inscribe-Mrc20
  /api/v1/market/address/assets:
    get:
      description: Fetch market address assets
      parameters:
      - description: pins/ordinals
        in: query
        name: assetType
        required: true
        type: string
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: default:0
        in: query
        name: cursor
        type: integer
      - description: default:10, max:50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.AddressAssetListResp'
      summary: Fetch market address assets
      tags:
      - Market-Asset
  /api/v1/market/asset/detail:
    get:
      description: Fetch market asset detail
      parameters:
      - description: assetId
        in: query
        name: assetId
        required: true
        type: string
      - description: pins/ordinals
        in: query
        name: assetType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.OrderInfo'
      summary: Fetch market asset detail
      tags:
      - Market-Asset
  /api/v1/market/mrc20/order/cancel:
    post:
      description: Cancel Market mrc20 order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.CancelMrc20OrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.CancelMrc20OrderResp'
      summary: Cancel Market mrc20 order
      tags:
      - Market-Order
  /api/v1/market/mrc20/order/detail:
    get:
      description: Fetch market mrc20 order detail
      parameters:
      - description: orderId
        in: query
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20OrderInfo'
      summary: Fetch market mrc20 order detail
      tags:
      - Market-Order
  /api/v1/market/mrc20/order/psbt:
    get:
      description: Fetch market mrc20 order psbt
      parameters:
      - description: orderId
        in: query
        name: orderId
        required: true
        type: string
      - description: buyerAddress
        in: query
        name: buyerAddress
        required: true
        type: string
      - description: buyerChangeAmount
        in: query
        name: buyerChangeAmount
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20OrderInfo'
      summary: Fetch market mrc20 order psbt
      tags:
      - Market-Order
  /api/v1/market/mrc20/order/push:
    post:
      description: Push Market mrc20 order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.PushMrc20OrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.PushMrc20OrderResp'
      summary: Push Market mrc20 order
      tags:
      - Market-Order
  /api/v1/market/mrc20/order/take:
    post:
      description: Take Market mrc20 order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.TakeMrc20OrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.TakeMrc20OrderResp'
      summary: Take Market mrc20 order
      tags:
      - Market-Order
  /api/v1/market/mrc20/orders:
    get:
      description: Fetch market mrc20 orders
      parameters:
      - description: mrc20
        in: query
        name: assetType
        required: true
        type: string
      - description: 1-create, 2-cancel, 3-finish
        in: query
        name: orderState
        required: true
        type: integer
      - description: tickId
        in: query
        name: tickId
        type: string
      - description: address
        in: query
        name: address
        type: string
      - description: priceAmount/timestamp/tokenPriceRate, default:timestamp
        in: query
        name: sortKey
        type: string
      - description: -1/1, default:-1
        in: query
        name: sortType
        type: integer
      - description: default:0
        in: query
        name: cursor
        type: integer
      - description: default:10, max:50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20OrderListResp'
      summary: Fetch market mrc20 orders
      tags:
      - Market-Order
  /api/v1/market/order/cancel:
    post:
      description: Cancel Market order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.CancelOrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.CancelOrderResp'
      summary: Cancel Market order
      tags:
      - Market-Order
  /api/v1/market/order/detail:
    get:
      description: Fetch market order detail
      parameters:
      - description: orderId
        in: query
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.OrderInfo'
      summary: Fetch market order detail
      tags:
      - Market-Order
  /api/v1/market/order/psbt:
    get:
      description: Fetch market order psbt
      parameters:
      - description: orderId
        in: query
        name: orderId
        required: true
        type: string
      - description: buyerAddress
        in: query
        name: buyerAddress
        required: true
        type: string
      - description: buyerChangeAmount
        in: query
        name: buyerChangeAmount
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.OrderInfo'
      summary: Fetch market order psbt
      tags:
      - Market-Order
  /api/v1/market/order/psbt/batch:
    post:
      description: Fetch market order psbt
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.BatchFetchOrderPsbtReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.OrderInfo'
      summary: Fetch market order psbt
      tags:
      - Market-Order
  /api/v1/market/order/push:
    post:
      description: Push Market order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.PushOrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.PushOrderResp'
      summary: Push Market order
      tags:
      - Market-Order
  /api/v1/market/order/push/batch:
    post:
      description: Push Market order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.BatchPushOrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.BatchPushOrderResp'
      summary: Push Market order
      tags:
      - Market-Order
  /api/v1/market/order/take:
    post:
      description: Take Market order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.TakeOrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.TakeOrderResp'
      summary: Take Market order
      tags:
      - Market-Order
  /api/v1/market/order/take/batch:
    post:
      description: Take Market order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.BatchTakeOrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.TakeOrderResp'
      summary: Take Market order
      tags:
      - Market-Order
  /api/v1/market/orders:
    get:
      description: Fetch market orders
      parameters:
      - description: pins/ordinals
        in: query
        name: assetType
        required: true
        type: string
      - description: 1-create, 2-cancel, 3-finish
        in: query
        name: orderState
        required: true
        type: integer
      - description: address
        in: query
        name: address
        type: string
      - description: sellPriceAmount/timestamp/assetLevel, default:timestamp
        in: query
        name: sortKey
        type: string
      - description: default:”
        in: query
        name: filter-path
        type: string
      - description: default:0
        in: query
        name: filter-level
        type: integer
      - description: default:”
        in: query
        name: filter-uncastTickId
        type: string
      - description: -1/1, default:-1
        in: query
        name: sortType
        type: integer
      - description: default:0
        in: query
        name: cursor
        type: integer
      - description: default:10, max:50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.OrderListResp'
      summary: Fetch market orders
      tags:
      - Market-Order
schemes:
- https
swagger: "2.0"